restructured file architecture, removed unused json files, updated repo to go 1.18
