+++ stateRead/state_read.go
@@ -1,4 +1,4 @@
-package main
+package stateRead
 
 import (
 	"fmt"
